Change to return int

+++ src/Google2FA.php
@@ -163,7 +163,7 @@ class Google2FA implements Google2FAContract
      **/
     public function getTimestamp()
     {
-        return floor(microtime(true) / $this->keyRegeneration);
+        return (int) floor(microtime(true) / $this->keyRegeneration);
     }
 
     /**
